move swiper pagination t styled components
